Fix: UserSelectionView 분리, 뷰모델 수정
