+++ Kabinett/Presentation/ViewModel/LetterWrite/UserSelectionViewModel.swift
@@ -10,35 +10,53 @@ import SwiftUI
 import Combine
 
 class UserSelectionViewModel: ObservableObject {
+    @Published var searchText: String = ""
+    @Published var checkLogin: Bool = false
+    @Published var userKabi: Int? = nil
+    @Published var fromUser: Writer? = nil
+    @Published var toUser: Writer? = nil
+    
     @Published var dummyUsers: [Writer] = [
-        Writer(name: "Alice", kabinettNumber: 111111, profileImage: "https://cdn.pixabay.com/photo/2022/06/25/13/33/landscape-7283516_1280.jpg"),
-        Writer(name: "Bob", kabinettNumber: 234234, profileImage: nil),
-        Writer(name: "Charlie", kabinettNumber: 111112, profileImage: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTpxjfCS04oKkgLWiCPCQg026DciIS5ayfvTg&s"),
-        Writer(name: "David", kabinettNumber: 111131, profileImage: nil),
-        Writer(name: "Eve", kabinettNumber: 1141, profileImage: nil),
-        Writer(name: "Frank", kabinettNumber: 111151, profileImage: nil),
-        Writer(name: "Grace", kabinettNumber: 111161, profileImage: nil),
+        Writer(id: "user_1", name: "Alice", kabinettNumber: 111111, profileImage: "https://cdn.pixabay.com/photo/2022/06/25/13/33/landscape-7283516_1280.jpg"),
+        Writer(id: "user_2", name: "Bob", kabinettNumber: 234234, profileImage: nil),
+        Writer(id: "user_3", name: "Charlie", kabinettNumber: 1112, profileImage: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTpxjfCS04oKkgLWiCPCQg026DciIS5ayfvTg&s"),
+        Writer(id: "user_4", name: "David", kabinettNumber: 11131, profileImage: nil),
+        Writer(id: "user_5", name: "Eve", kabinettNumber: 114141, profileImage: nil),
+        Writer(id: "user_6", name: "Frank", kabinettNumber: 1151, profileImage: nil),
+        Writer(id: "user_7", name: "Grace", kabinettNumber: 111161, profileImage: nil),
     ]
     
-    @Published var loginUserKabinett: Int? = 111111
-    @Published var checkLogin: Bool = false
-    
-    @Published var fromUser: String = ""
-    @Published var toUser: String = ""
-    
     init() {
+        userKabi = nil
         updateFromUser()
     }
     
     private func updateFromUser() {
-        if let user = dummyUsers.first(where: { $0.kabinettNumber == loginUserKabinett }) {
+        if let user = dummyUsers.first(where: { $0.kabinettNumber == userKabi }) {
             checkLogin = true
-            fromUser = user.name + "(나)"
-            toUser = user.name + "(나)"
+            fromUser = Writer(id: user.id, name: user.name, kabinettNumber: user.kabinettNumber, profileImage: user.profileImage)
         } else {
             checkLogin = false
-            fromUser = "나"
-            toUser = "나"
+            fromUser = Writer(name: "나", kabinettNumber: 0, profileImage: nil)
+            toUser = Writer(name: "나", kabinettNumber: 0, profileImage: nil)
         }
+    }
+    
+    func updateToUser(_ letterContent: inout LetterWriteViewModel, toUserName: String) {
+        if let user = dummyUsers.first(where: { $0.name == toUserName }) {
+            toUser = Writer(id: user.id, name: user.name, kabinettNumber: user.kabinettNumber, profileImage: user.profileImage)
+            letterContent.toUserId = toUser?.id
+            letterContent.toUserName = toUser?.name ?? ""
+            letterContent.toUserKabinettNumber = toUser?.kabinettNumber
+        } else {
+            toUser = Writer(name: toUserName, kabinettNumber: 0, profileImage: nil)
+            letterContent.toUserId = toUser?.id
+            letterContent.toUserName = toUser?.name ?? ""
+            letterContent.toUserKabinettNumber = toUser?.kabinettNumber
+        }
+    }
+    
+    func checkMe(kabi: Int) -> String {
+        userKabi == kabi ? "(나)" : ""
     }
 }
